requirements.txt
anthropic==0.40.0
tavily-python==0.5.0
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.3
aiohttp==3.11.10